[DT-51] Moved widgetsService to app module
