+++ app/src/main/java/cm/aptoide/pt/apps/WidgetsNetworkService.kt
@@ -1,13 +1,14 @@
-package cm.aptoide.pt.feature_apps.data.network.service
+package cm.aptoide.pt.apps
 
 import cm.aptoide.pt.feature_apps.data.network.model.BaseV7DataListResponse
 import cm.aptoide.pt.feature_apps.data.network.model.WidgetsJSON
+import cm.aptoide.pt.feature_apps.data.network.service.WidgetsRemoteService
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
 import javax.inject.Inject
 
-internal class WidgetsNetworkService @Inject constructor(
+class WidgetsNetworkService @Inject constructor(
   private val widgetsRemoteDataSource: Retrofit,
   private val storeName: String
 ) :
@@ -17,7 +18,7 @@ internal class WidgetsNetworkService @Inject constructor(
     return widgetsRemoteDataSource.getStoreWidgets(storeName)
   }
 
-  internal interface Retrofit {
+  interface Retrofit {
     @GET("getStoreWidgets?aptoide_vercode=20000&limit=25")
     suspend fun getStoreWidgets(
       @Query("store_name") storeName: String
