+++ app/src/main/java/cm/aptoide/pt/di/RepositoryModule.kt
@@ -1,12 +1,17 @@
 package cm.aptoide.pt.di
 
 import cm.aptoide.pt.BuildConfig
+import cm.aptoide.pt.apps.WidgetsNetworkService
+import cm.aptoide.pt.aptoide_network.di.RetrofitV7
 import cm.aptoide.pt.aptoide_network.di.StoreName
+import cm.aptoide.pt.feature_apps.data.di.WidgetsService
+import cm.aptoide.pt.feature_apps.data.network.service.WidgetsRemoteService
 import cm.aptoide.pt.home.BottomNavigationManager
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
 import javax.inject.Singleton
 
 @Module
@@ -21,4 +26,15 @@ class RepositoryModule {
   @Provides
   @StoreName
   fun provideStoreName(): String = BuildConfig.MARKET_NAME
+
+  @Provides
+  @Singleton
+  @WidgetsService
+  fun providesWidgetsRemoteService(
+    @RetrofitV7 retrofitV7: Retrofit,
+    @StoreName storeName: String
+  ): WidgetsRemoteService = WidgetsNetworkService(
+    widgetsRemoteDataSource = retrofitV7.create(WidgetsNetworkService.Retrofit::class.java),
+    storeName = storeName
+  )
 }
